--- WebServer/Base/__init__.py
@@ -1,1 +1,0 @@
-from .ServerBase import *
