--- WebServer/Configurations/Settings/__init__.pyi
@@ -1,1 +1,0 @@
-from .Settings import *
