--- WebServer/Utils/utils/__init__.py
@@ -1,3 +1,0 @@
-from .render_template import *
-from .utils import *
-from .content_type import *
