+++ WebServer/__init__.py
@@ -1,3 +1,3 @@
 from .core.WebServer import *
-from .Utils.Request import *
-from .Utils.Response import *
+from .http.Request import *
+from .http.Response import *
