+++ WebServer/config/settings/__init__.pyi
@@ -1,0 +1,1 @@
+from .settings import *
