+++ WebServer/http/Request/Request.pyi
@@ -1,7 +1,5 @@
-from ...types import RequestType
-from typing import TypedDict, Literal, Union, Dict
+from WebServer.types import TypedDict, Literal, Union, Dict, RequestMethod
 from dataclasses import dataclass, field
-
 
 _Headers = TypedDict(
     "_Headers",
@@ -24,12 +22,10 @@ _Headers = TypedDict(
     },
 )
 
-
 @dataclass
 class Request:
     url: str
 
-    # ? Dynamic Fields
     path: str = field(init=False)
     scheme: str = field(init=False)
     netloc: str = field(init=False)
@@ -40,7 +36,7 @@ class Request:
     Host: str
     Connection: Literal["keep-alive", "close"]
     headers: Union[_Headers, Dict[str, str]]
-    method: RequestType
+    method: RequestMethod
 
     user_parameters: Dict[str, str] | None = field(init=False, default=None)
