+++ requirements.txt
@@ -1,0 +1,6 @@
+# DevDependencies
+setuptools
+djlint
+
+# Dependencies
+colorama
